requirements.txt
absl-py==2.1.0
altair==5.5.0
aniso8601==10.0.1
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
av==14.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.1.1
Flask==3.1.0
Flask-RESTful==0.3.10
flatbuffers==20181003210633
fonttools==4.58.0
future==1.0.0
gast==0.5.4
geocoder==1.38.1
geographiclib==2.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.7
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==2.15.0
kiwisolver==1.4.8
lazy_loader==0.4
libarchive-c==5.2
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.6
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.9
ml-dtypes==0.2.0
msgpack==1.0.8
narwhals==1.37.1
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
osmnx==2.0.2
packaging==24.0
pandas==2.2.3
pidng==4.0.9
piexif==1.1.3
pillow==11.2.1
platformdirs==4.2.1
polyline==2.0.2
pooch==1.8.1
protobuf==3.20.3
pyarrow==20.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydeck==0.9.1
pyfiglet==1.0.2
pynmea2==1.19.0
pynmeagps==1.0.49
pyogrio==0.10.0
pyparsing==3.2.3
pyperclip==1.9.0
pyproj==3.7.1
pyrtcm==1.1.6
pyserial==3.5
python-dateutil==2.9.0.post0
python-prctl==1.8.1
pytz==2025.2
pyubx2==1.2.51
PyYAML==6.0.2
ratelim==0.1.6
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
RPi.GPIO==0.7.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
shapely==2.1.0
simplejpeg==1.8.2
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.3.7
streamlit==1.45.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-cpu-aws==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
tkintermapview==1.29
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
v4l2-python3==0.3.5
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.14.1